Make structural changes window MVVM
- fix bugs in Variable Comparison
- move structural comparison logic to separate class
