--- Controller/Controller/Control/SimpleStringOkWindow.xaml
@@ -1,16 +1,0 @@
-﻿<Window
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="d" x:Class="Controller.Control.SimpleStringOkWindow"
-        Title="SimpleStringOkWindow" Height="600" Width="800">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="25"/>
-        </Grid.RowDefinitions>
-        <ScrollViewer Grid.Row="0">
-            <TextBlock x:Name="TextBlock1" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"/>
-        </ScrollViewer>
-        <Button Grid.Row="1" Click="ButtonBase_OnClick" Content="OK"/>
-    </Grid>
-</Window>
