+++ Controller/Controller/Helper/FileHelper.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.IO.Compression;
 using System.Runtime.Serialization;
 
-namespace Controller
+namespace Controller.Helper
 {
     public class FileHelper
     {
