+++ Controller/Controller/MainWindow/MeasurementRoutine/RoutineModelController.cs
@@ -3,6 +3,8 @@ using Model.Root;
 using System.Windows.Input;
 using Communication.Commands;
 using System.Windows;
+using Controller.Common;
+using Controller.Helper;
 
 namespace Controller.MainWindow.MeasurementRoutine
 {
@@ -81,7 +83,7 @@ namespace Controller.MainWindow.MeasurementRoutine
         public void LoadPythonScripts(object parameter)
         {
             PythonScriptsController pyhtonScriptsController = new PythonScriptsController(this.RoutineModel);
-            Window pyhtonScriptsWindow = WindowsHelper.CreateCustomWindowToHostViewModel(pyhtonScriptsController, false);
+            Window pyhtonScriptsWindow = WindowsHelper.CreateWindowToHostViewModel<CustomWindow>(pyhtonScriptsController, false);
 
             pyhtonScriptsWindow.MinHeight = 360;
             pyhtonScriptsWindow.MinWidth = 550;
