+++ View/View/Control/Compare/SimpleMessageCompareItem.xaml
@@ -1,17 +1,16 @@
 ﻿<UserControl
-    x:Class="View.Control.WindowsTemplate"
+    x:Class="View.Control.Compare.SimpleMessageCompareItem"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    d:DesignHeight="300"
-    d:DesignWidth="300"
+    d:DesignHeight="450"
+    d:DesignWidth="800"
     mc:Ignorable="d">
-    <Grid>
-        <Button
-            Name="TheName"
-            Grid.Column="0"
-            Command="{Binding SelectWindow}"
-            Content="{Binding Name, Mode=OneWay}" />
+    <Grid Margin="10">
+        <TextBlock
+            Text="{Binding Message}"
+            TextAlignment="Left"
+            TextWrapping="Wrap" />
     </Grid>
 </UserControl>
