+++ View/View/Control/SimpleStringOkDialog.xaml
@@ -1,0 +1,33 @@
+﻿<UserControl
+    x:Class="View.Control.SimpleStringOkDialog"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    x:Name="UC"
+    MinWidth="400"
+    MinHeight="300"
+    mc:Ignorable="d">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*" />
+            <RowDefinition Height="auto" />
+        </Grid.RowDefinitions>
+        <ScrollViewer Grid.Row="0">
+            <TextBlock
+                Margin="10"
+                HorizontalAlignment="Stretch"
+                VerticalAlignment="Stretch"
+                Text="{Binding Message}" />
+        </ScrollViewer>
+        <Button
+            Grid.Row="1"
+            Width="100"
+            Height="30"
+            Margin="10"
+            HorizontalAlignment="Stretch"
+            Command="{Binding OkClickedCommand}"
+            CommandParameter="{Binding ElementName=UC}"
+            Content="OK" />
+    </Grid>
+</UserControl>
