+++ View/View/Control/SimpleStringOkDialog.xaml.cs
@@ -1,0 +1,13 @@
+﻿using System.Windows;
+using System.Windows.Controls;
+
+namespace View.Control
+{
+    public partial class SimpleStringOkDialog : UserControl
+    {
+        public SimpleStringOkDialog()
+        {
+            InitializeComponent();
+        }
+    }
+}
